backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Google Gemini AI
google-generativeai>=0.3.0

# Supabase client
supabase>=2.0.0

# Pydantic for data validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# HTTP client
httpx>=0.26.0

# JWT authentication
PyJWT>=2.8.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.0.0

# Development
black>=24.0.0
ruff>=0.1.0
mypy>=1.0.0
